# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "应用于快速搭建后端",
        "title": "基础后端框架",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "http://127.0.0.1:8001",
    "basePath": "/v1/",
    "paths": {
...
        "/user/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户创建",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "用户创建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Data"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.Create"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/:user_id/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "查询指定用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "查询指定用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Data"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改指定用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "修改指定用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserUpdateById"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Data"
                        }
                    }
                }
            }
        },
        "/user/avatar/": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户头像修改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "用户头像修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "头像",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Data"
                        }
                    }
                }
            }
        },
        "/user/change_pwd/": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户密码修改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "用户密码修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户密码信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserChangePwdBase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Data"
                        }
                    }
                }
            }
        },
        "/user/change_status/:user_id/": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改指定账户状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "修改指定账户状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户状态",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.StatusChangeById"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Data"
                        }
                    }
                }
            }
        },
        "/user/detail/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "用户详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Data"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户修改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "用户修改",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户修改信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Data"
                        }
                    }
                }
            }
        },
        "/user/list/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "x-enum-varnames": [
                            "AccNormal",
                            "AccFreeze",
                            "AccChangePwd",
                            "AccStop"
                        ],
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Data"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserList"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/login/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "登录参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserLoginBase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Data"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.LoginSuccess"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/logout/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "登出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "登出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Data"
                        }
                    }
                }
            }
        },
        "/user/reset_pwd/:user_id/": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "重置指定账号密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserApi"
                ],
                "summary": "重置指定账号密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 令牌",
                        "name": "Identification",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户密码信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.PwdChangeById"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Data"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.StatusChangeById": {
            "type": "object",
            "properties": {
                "status": {
                    "$ref": "#/definitions/user.AccStatus"
                }
            }
        },
...
        "request.UserCreate": {
            "type": "object",
            "required": [
                "account",
                "password"
            ],
            "properties": {
                "account": {
                    "type": "string",
                    "maxLength": 20
                },
                "createTime": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/gender.Gender"
                },
                "id": {
                    "type": "integer"
                },
                "id_card": {
                    "type": "string",
                    "maxLength": 18
                },
                "name": {
                    "type": "string",
                    "maxLength": 20
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 11
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "secret_key": {
                    "type": "string"
                },
                "short_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/user.AccStatus"
                },
                "updateTime": {
                    "type": "integer"
                }
            }
        },
...
        "request.UserUpdateById": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createTime": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/gender.Gender"
                },
                "id_card": {
                    "type": "string",
                    "maxLength": 18
                },
                "name": {
                    "type": "string",
                    "maxLength": 20
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "short_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/user.AccStatus"
                },
                "updateTime": {
                    "type": "integer"
                }
            }
        },
        "response.Create": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "response.Data": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "integer"
                },
                "status_info": {
                    "$ref": "#/definitions/response.statusInfo"
                }
            }
        },
...
        "response.UserDetail": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "create_time": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "id_card": {
                    "type": "string"
                },
                "is_system": {
                    "type": "boolean"
                },
                "last_time": {
                    "type": "integer"
                },
                "login_type": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "privilege_list": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "$ref": "#/definitions/user.AccStatus"
                }
            }
        },
...
    }
}